BUGFIX: Cleaned up the creation and saving of textgrids and tiers
Before, textgrids could be opened with silences removed (which would
make a textgrid with no entries unopenable) and textgrids could be saved
with 'gaps' in the textgrid (which would lead to strange behavior
in praat, such as having unselectable regions of the textgrid).
Both of these possibilities have been removed.