BUGFIX: Changed kgio to store values as decimal rather than floats
In practice, it doesn't make sense to have floats so large but
praat is very precise with its time estimates which is stored to
a more precise level than float.

But this means some converting will need to be done by whatever
tools process the data. (Cast to float, for example)